container: move custom handlers code to new file
