+++ src/crun.c
@@ -27,6 +27,7 @@
 
 #include "crun.h"
 #include "libcrun/utils.h"
+#include "libcrun/custom-handler.h"
 
 /* Commands.  */
 #include "run.h"
@@ -207,14 +208,9 @@ print_version (FILE *stream, struct argp_state *state arg_unused)
 #ifdef HAVE_CRIU
   fprintf (stream, "+CRIU ");
 #endif
-#ifdef HAVE_WASMER
-  fprintf (stream, "+WASM:wasmer ");
-#elif defined HAVE_WASMEDGE
-  fprintf (stream, "+WASM:wasmedge ");
-#endif
-#ifdef HAVE_LIBKRUN
-  fprintf (stream, "+LIBKRUN ");
-#endif
+
+  print_handlers_feature_tags (stream);
+
   fprintf (stream, "+YAJL\n");
 }
